Change the Category Icon to a Search Icon and A Few Thoughts
After asking several friends, I found people are confusing about the
original category icon, so I decided to move the search to the category.
The reason behind this is that if users cannot find the proper live
program on the home page, then they would like to search for the
program. And instead of only proving a search bar, showing category
with recommendations would be more intuitive for users to choose. The
action of finding a category is also an action of searching.
I think this change would better help users to find what they want in
the app.

Also, I think it would be better to just have a single page for the
four sections of the main page. On an iOS app, users have the intention
to swipe right to see what was going on and when they came back, it
would be better to have the original position on the page as well as
the content.
For example, after many levels of finding in the search section, a user
wants to see his collection. While he came back to the search section,
it would be a lot convenient for him to have the page that has already
been loaded before.

THINGS TO ADD:
A category of each program.

THINGS SHOULD BE CONSIDERED TO CHANGE:
The UI of the main page is okay, because it would be hard to insert a
lot more useful information. I need to find a way to change the UI a
little bit, or maybe a big overhaul? Right now I have the idea of add a
“more” to the description section, which users can tap on and see more
info about the program.